+++ samples/CreateDebugUtilsMessenger/CreateDebugUtilsMessenger.cpp
@@ -60,7 +60,7 @@ VKAPI_ATTR VkBool32 VKAPI_CALL debugMessageFunc( VkDebugUtilsMessageSeverityFlag
   {
     message << "\t"
             << "Queue Labels:\n";
-    for ( uint8_t i = 0; i < pCallbackData->queueLabelCount; i++ )
+    for ( uint32_t i = 0; i < pCallbackData->queueLabelCount; i++ )
     {
       message << "\t\t"
               << "labelName = <" << pCallbackData->pQueueLabels[i].pLabelName << ">\n";
@@ -70,7 +70,7 @@ VKAPI_ATTR VkBool32 VKAPI_CALL debugMessageFunc( VkDebugUtilsMessageSeverityFlag
   {
     message << "\t"
             << "CommandBuffer Labels:\n";
-    for ( uint8_t i = 0; i < pCallbackData->cmdBufLabelCount; i++ )
+    for ( uint32_t i = 0; i < pCallbackData->cmdBufLabelCount; i++ )
     {
       message << "\t\t"
               << "labelName = <" << pCallbackData->pCmdBufLabels[i].pLabelName << ">\n";
@@ -80,7 +80,7 @@ VKAPI_ATTR VkBool32 VKAPI_CALL debugMessageFunc( VkDebugUtilsMessageSeverityFlag
   {
     message << "\t"
             << "Objects:\n";
-    for ( uint8_t i = 0; i < pCallbackData->objectCount; i++ )
+    for ( uint32_t i = 0; i < pCallbackData->objectCount; i++ )
     {
       message << "\t\t"
               << "Object " << i << "\n";
